Multiplicity of infection in Mali using 24-SNP barcode data (Adomako-Ankomah et al., 2017)

Data from Adomako-Ankomah et al. (2017). Here we give a brief summary of the data - see the original paper for full details.

A longitudinal cohort study was conducted in a seasonal and high-transmission area of Mali, in which 500 subjects aged 1–65 years were followed for 1 year. Blood samples were collected every 2 weeks. Multiclonality of Pf infection was measured using a 24-SNP DNA barcoding assay at 4 time-points (two in wet season, and two in dry season).

data(Adomako_Ankomah_2017)

Format

A dataframe with 26 columns, giving time point in the year (column 1), multiplicity category, including number of failed assays (column 2), and frequencies for each of these classes for all 24 SNPs (columns 3:26).

Source

Supplementary materials

References

Adomako-Ankomah Y, Chenoweth MS, Durfee K, Doumbia S, Konate D, Doumbouya M, Keita AS, Nikolaeva D, Tullo GS, Anderson JM, Fairhurst RM, Daniels R, Volkman SK, Diakite M, Miura K, Long CA (2017). “High Plasmodium falciparum longitudinal prevalence is associated with high multiclonality and reduced clinical malaria risk in a seasonal transmission area of Mali.” PLOS ONE, 12(2), e0170948. ISSN 1932-6203, doi: 10.1371/journal.pone.0170948 , https://dx.plos.org/10.1371/journal.pone.0170948.