24-SNP barcode data at two time-points in Senegal (Bei et al., 2018)

Data from Bei et al. (2018). Here we give a brief summary of the data - see the original paper for full details.

Samples were randomly selected from longitudinal cohorts collected from Dielmo/Ndiop, Senegal, and focus on 2 distant time-points (2001-2002, and 2014). The first time-point corresponds to a period of high transmission, and the second a period of extremely low transmission. Samples were genotyped using a 24-SNP barcode, and complexity of infection was estimated using the COIL algorithm. Samples were also matched against a large database of previously published and unpublished barcodes. For samples from the low transmission period (2014), one of the three repeated barcode clusters (n = 6) corresponded to a parasite type (haplotype 3), observed in Thiès in both 2007 and 2010. The other 2 clusters (IP1, n = 2; IP2, n = 2) had not previously been observed. Data were extracted using Tabula v1.2.1.

data(Bei_2018)

Format

A list of multiple data objects:

EIR: estimates of the EIR at each location and year

barcodes: genetic data and associated sample characteristics

SNP_locations: genomic locations of SNPs (key to be used with barcodes)

EIR: A dataframe with 3 columns, giving the time, the sampling location, and the estimated EIR (see original paper for details). The EIR in 2014 in Ndiop was recorded as "0.0" in the paper, and so has been coded as "<0.05" here to indicate the precision of this estimate.

barcodes: A dataframe with 34 columns. Gives sample characteristics (columns 1:7), the estimated COI and whether this indicates a monogenomic/polygenomic infection (columns 8:9), the individual SNPs (columns 10:33) and the corresponding haplotype, if known (column 34).

SNP_locations: A dataframe that acts as a key relating the SNP codes present in barcodes to the corresponding genomic location.

Source

Data in main paper Tabula data extractor

References

Bei AK, Niang M, Deme AB, Daniels RF, Sarr FD, Sokhna C, Talla C, Faye J, Diagne N, Doucoure S, Mboup S, Wirth DF, Tall A, Ndiaye D, Hartl DL, Volkman SK, Toure-Balde A (2018). “Dramatic Changes in Malaria Population Genetic Complexity in Dielmo and Ndiop, Senegal, Revealed Using Genomic Surveillance.” The Journal of Infectious Diseases, 217(4), 622--627. ISSN 0022-1899, 1537-6613, doi: 10.1093/infdis/jix580 , https://academic.oup.com/jid/article/217/4/622/4793403.