Inferring local and crossborder transmission with human and mobility data, Kanunga District in Southwest Uganda (Briggs et al., 2019)

Data from Briggs et al. (2021). Here we give a brief summary of the data - see the original paper for full details.

The study uses 80 randomly sampled households around a single health facility and includes all children (6 mo - 10 years old) and at least one adult caretaker from the households. DNA was acquired though dried blood spots, and after two rounds of PCR amplification, the PCR products were sized using capillary electrophoresis. Finally, allele length was calculated.

data(Briggs_2021)

Format

A list of multiple data objects:

final_samples_Dec: contains allele lengths with barcodes.

Kanungu_pairwise_comparison_df: epidemiological data set with final_samples_Dec keys.

Kanungu_pairwise_comparison_df: Metadata for the allele sizes located in final_samples_Dec.

Source

Supplementary file 1

References

Briggs J (2021). “Within‐household clustering of genetically related Plasmodium falciparum infections in a moderate transmission area of Uganda.” Malaria journal, 8.