24-SNP barcode data showing population structure in Haiti (Charles et al., 2016)

Data from Charles et al. (2016). Here we give a brief summary of the data - see the original paper for full details.

Samples were collected through routine surveillance spanning 2006–2009 by the Haitian Group for the Study of Kaposi’s Sarcoma and Opportunistic Infections (GHESKIO) at 9 healthcare centers in various municipalities. Samples were genotyped using a 24-SNP barcode. The paper states that samples that showed >1 mixed-base SNP call or had >5 missing calls in the 24SNP molecular barcode were removed from analysis, however, there is 1 sample (monogenomic sample 24) that has 6 missing calls and has not been removed, and there is also 1 sample (polygenomic sample 8) that has exactly 1 mixed-base SNP call and 0 missing calls and yet has been removed. Therefore, it is possible that samples were removed if they showed >0 mixed-base calls, or had >6 missing calls (all data presented are consistent with this filtering). Both monogenomic (n = 52) and polygenomic (n = 8) samples are available, although samples excluded based on missingness are not available. Samples were compared in terms of the proportion of shared bases, and were identified as either identical, related, or unique based on this number.

data(Charles_2016)

Format

A list of two data objects:

monoclonal: the 42 samples identified as monoclonal

polyclonal: the remaining 8 samples identified as polyclonal

monoclonal: A dataframe with 6 columns, giving sample characteristics (columns 1:3), barcode data (columns 4:5), and similarity category (column 6). location, and the estimated EIR (see original paper for details). Heterozygous genotyping calls are identified by "N", and missing alleles are identified by "X".

polyclonal: A dataframe with 4 columns, giving sample characteristics (columns 1:3), and barcode data (column 4). As above, heterozygous genotyping calls are identified by "N", and missing alleles are identified by "X".

Source

Data extracted from tables

References

Charles M, Das S, Daniels R, Kirkman L, Delva GG, Destine R, Escalante A, Villegas L, Daniels NM, Shigyo K, Volkman SK, Pape JW, Golightly LM (2016). “Plasmodium falciparum K76T pfcrt Gene Mutations and Parasite Population Structure, Haiti, 2006–2009.” Emerging Infectious Diseases, 22(5), 786--793. ISSN 1080-6040, 1080-6059, doi: 10.3201/eid2205.150359 , http://wwwnc.cdc.gov/eid/article/22/5/15-0359_article.htm.