24-SNP barcode data showing clonal and epidemic transmission in Senegal (Daniels et al., 2013)

Data from Daniels et al. (2013). Here we give a brief summary of the data - see the original paper for full details.

Note - see ?Galinsky_2015 for an expanded version of this same set of barcodes. All 529 Sample IDs in this dataset are also present in Galinsky_2015, along with an additional 445 samples.

Samples were obtained annually from 2006–2011 from the Service de Lutte Anti-Parasitaire (SLAP) clinic in Thies, Senegal. Samples were collected passively; with patients over the age of 12 months admitted to this study with self-reported acute fevers within 24 hours of visiting the clinic and no recent anti-malarial use. Whole blood spots from 2006–2011 were preserved on filter paper, and DNA was extracted and characterised using a 24-SNP molecular barcode. Samples were excluded if they had missing data on more than four SNP positions. Samples were identified as polygenomic if they had multiallelic calls at more than one SNP position, otherwise monogenomic. It is unclear whether the data presented is for monogenomic samples only, as one sample (SenT081) has mixed calls at 2 sites and therefore is polygenomic, but it seems unlikely that there was only one polygenomic sample observed in the entire dataset.

data(Daniels_2013)

Format

A dataframe with 27 columns, giving sample characteristics (columns 1:3), and genomic data at 24 SNPs (columns 4:27). Heterozygous genotyping calls are identified by "N", and missing alleles are identified by "NA".

Source

Supplementary materials

References

Daniels R, Chang H, Séne PD, Park DC, Neafsey DE, Schaffner SF, Hamilton EJ, Lukens AK, Tyne DV, Mboup S, Sabeti PC, Ndiaye D, Wirth DF, Hartl DL, Volkman SK (2013). “Genetic Surveillance Detects Both Clonal and Epidemic Transmission of Malaria following Enhanced Intervention in Senegal.” PLOS ONE, 8(4), e60780. ISSN 1932-6203, doi: 10.1371/journal.pone.0060780 , Publisher: Public Library of Science, https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0060780.