Measures of prevalence by PCR and Microscopy

Data on Plasmodium falciparum prevalence by microscopy vs. by PCR. Taken from the systematic review by Whittaker et al. (Lancet Microbe, 2021), downloaded from the associated GitHub repository. This dataset contains prevalence in all ages - see ?PCR_micro_age_whittaker2021 for values broken down by age. Note, the original repository contains more information about where these values come from, and also more columns that were not considered relevant for our purposes.

data(PCR_micro_full_whittaker2021)

Format

A data frame of 387 rows and 19 columns. Columns give the name and year of the study, the study location at multiple geographic levels, measures of historical and current transmission intensity in terms of regional prevalence, the numerator, denominator and resulting prevalence by both PCR and microscopy, details of the PCR and microscopy methods, and finally the sampling season. Some fields are given in terms of "Raw" values in addition to processed values. Raw values have more flexibility and are closer to the description in the original papers, while processed values are simplified to a smaller number of possible levels.

Source

Systematic Review Results

References

Charles Whittaker, Hannah Slater, Rebecca Nash, Teun Bousema, Chris Drakeley, Azra C. Ghani, and Lucy C. Okell. Global patterns of submicroscopic Plasmodium falciparum malaria infection: insights from a systematic review and meta-analysis of population surveys. Lancet Microbe (2021) https://doi.org/10.1016/S2666-5247(21)00055-0 (ScienceDirect)