Microsatellite data across 26 loci, Eswatini (Roh et al., 2019)

Data from Roh et al. (2019). Here we give a brief summary of the data - see the original paper for full details.

The study investigates local versus imported transmissions and how that effects the genetic diversity in Eswatini. The parameters collected in the epidemiological model include district of residence, season, case detection, age, sex, and patient occupation. To compare genetic diversity, an overlapping subset of microsatellite markers were used to compare against other countries with lower transmission rates (it was found that Eswatini had a greater diversity of parasites (mean $H_E$ was consistently higher)). When comparing the local and imported cases, population level diversity did not differentiate greatly; however, a greater portion of local cases were monoclonal and less complex. Differences in MOI and $F_ws$ between cases were statistically significant but had marginal effect sizes and could not be used to discriminate cases.

Eligible cases were identified through Eswatini's national malaria surveillance program. Samples were genotyped using microsatellites from 26 different loci, which showed no selection based on $F_ws$ calculations. In their study, they excluded data points where the probability of there being an allele was less than 95

data(Roh_2019)

Format

A data frame with 10 columns, giving sample ID and barcode (columns 1:2), locus, locus indexes, and a combined id (columns 3:5), peaks and fluorescence (columns 6:9), and probability of being a true allele (column 10).

Source

Supplementary materials

References

Roh ME, Tessema SK, Murphy M, Nhlabathi N, Mkhonta N, Vilakati S, Ntshalintshali N, Saini M, Maphalala G, Chen A, Wilheim J, Prach L, Gosling R, Kunene S, S. Hsiang M, Greenhouse B (2019). “High Genetic Diversity of Plasmodium falciparum in the Low-Transmission Setting of the Kingdom of Eswatini.” The Journal of Infectious Diseases, 220(8), 1346--1354. ISSN 0022-1899, 1537-6613, doi: 10.1093/infdis/jiz305 , https://academic.oup.com/jid/article/220/8/1346/5514501.