Microsatellite data across 26 loci, Northeast Namibia (Tessema et al., 2019)

Data from Tessema et al. (2019). Here we give a brief summary of the data - see the original paper for full details.

The study collected samples from 4643 symptomatic outpatients in northeast Namibia (Kavango East and Zambezi) with confirmed cases via rapid tests. 3871 cases were from Kavango East (March-June 2016) from 23 different health facilities, and 772 from Zambezi (Feb 2015 and June 2016) from six facilities.

DNA was extracted from dried blood spots using punches and strips from rapid diagnostic tests. If a facility had less than 100 rapid tests, all tests were included for genotyping. If over 100 tests, then all cases with travel history were included for genotyping, and up to 100 cases without history were also included. In total, 2990 samples were genotyped using microsatellite markers, in which 2585 were used for data analysis.

data(Tessema_2019)

Format

A data frame with 31 columns, giving case ID, Health Facility, District, Region, Country (Columns 1:5), and microsatellite lengths (6:31).

Source

Supplementary file 1

References

Tessema S, Wesolowski A, Chen A, Murphy M, Wilheim J, Mupiri A, Ruktanonchai NW, Alegana VA, Tatem AJ, Tambo M, Didier B, Cohen JM, Bennett A, Sturrock HJ, Gosling R, Hsiang MS, Smith DL, Mumbengegwi DR, Smith JL, Greenhouse B (2019). “Using parasite genetic and human mobility data to infer local and cross-border malaria connectivity in Southern Africa.” eLife, 8, e43510. ISSN 2050-084X, doi: 10.7554/eLife.43510 , Publisher: eLife Sciences Publications, Ltd, 2022-04-25.