Data on prevalence and incidence broken down by age, used in Griffin et al. 2014 model fit

Data used to fit the model described in Griffin et al. (2010) and Griffin et al. (2014). Data consists of multiple prevalence and incidence estimates, broken down by age and study site. Alongside observation counts (i.e. numerator and denominator values) data contains estimated distributions of EIR and treatment rates that can be used to define priors in model fitting.

prev_inc_griffin2014

Format

A dataframe of 262 rows and 15 columns. Columns are defined as follows:

country: country in which study was conducted.

reference: the original reference for this data. All references can be found in Griffin et al. (2010) and Griffin et al. (2014).

site_name: name of the study site.

site_index: a unique numerical index given to each site (sites are contained within studies).

age0, age1: lower and upper ends of this age group. Brackets are open on the right, for example if age0 = 1 and age1 = 2 then this corresponds to individuals of one year of age (rather than one and two year olds).

type: whether incidence or prevalence estimate.

numer, denom: numerator and denominator values. If prevalence data then denominator is total population size, if incidence data then denominator is total time at risk.

case_detection: method of case detection, for example active case detection (ACD) vs. passive case detection (PCD).

meanEIR, sd_hi, sd_low: estimate of the EIR in the region. Estimates are given in terms of a mean estimate and a lower and upper 95

alpha(prop treated), beta(prop treated): shape parameters of a Beta prior on treatment rates in the population. These can be used as a prior in model fitting, or the value alpha / (alpha + beta) can be used as a point estimate of the proportion of cases receiving treatment.

@keywords datasets

Details

@docType data

@usage data(prev_inc_griffin2014)

References

Jamie T. Griffin, Deirdra Hollingsworth, Lucy C. Okell, Thomas S. Churcher, Michael White, Wes Hinsley, Teun Bousema, Chris J. Drakeley, Neil M. Ferguson, Maria-Gloria Basanes and Azra C. Ghani. Reducing Plasmodium falciparum Malaria Transmission in Africa: A Model-Based Evaluation of Intervention Strategies. PLoS Medicine (2010) doi:10.1371/journal.pmed.1000324 (PubMed)

Jamie T. Griffin, Neil M. Ferguson and Azra C. Ghani. Estimates of the changing age-burden of Plasmodium falciparum malaria disease in sub-Saharan Africa. Nature Communications (2014) DOI: 10.1038/ncomms4136 (PubMed)