Retrieve prevalence

Retrieve prevalence

retrieve_prev(
  data,
  case_detection,
  diagnosis,
  sampling_time,
  sampled = FALSE,
  deme = 1
)

Arguments

data
Model output from SIMPLEGEN sim_epi() function epi_output$daily_values

case_detection
Method of case detection, "Active" or "Passive"

diagnosis
Method of diagnosis, "Microscopy" or "PCR"

sampling_time
Sampling time, currently input should be a numeric day e.g. 100 would represent the 100th day of simulation.

sampled
TO DO - sample from bernoulli distribution to generate simulated numbers of cases detected on each day - done in SIMPLEGEN internally already

deme
deme of interest - currently assumes just a single deme analysed - i.e. can't pool results of multiple demes

Value

returns named vector : c("annual_EIR", "prevalence_true", "prevalence_detected")