Draw tree of genome-wide relatedness from pruned transmission record

Reads in the pruned transmission record from file and creates a new node for each inoculation ID. Nodes at time zero are initialised with a single haplotype, created de novo with a unique haplotype ID. In subsequent generations the nodes that are ancestral to the focal node are known from the pruned transmission record. The haplotypes from ancestral nodes are sampled at random, and brought together in pairs to produce oocysts. The recombinant products of these oocysts are then sampled down to produce a new generation of haplotype IDs for this node, along with the relative densities of each haplotype.

sim_relatedness(
  project,
  pruned_record_location = "",
  r = 1e-06,
  alpha = 1,
  oocyst_distribution = dpois(1:10, lambda = 2),
  hepatocyte_distribution = dpois(1:10, lambda = 5),
  contig_lengths = c(643292, 947102, 1060087, 1204112, 1343552, 1418244, 1501717,
    1419563, 1541723, 1687655, 2038337, 2271478, 2895605, 3291871),
  silent = FALSE
)

Arguments

project
a SIMPLEGEN project, as produced by the simplegen_project() function.

pruned_record_location
the file path from which the pruned transmission record will be read.

r
the rate of recombination. The expected number of base pairs in a single recombinant block is 1/r.

alpha
parameter dictating the skew of haplotype densities. Small values of alpha create a large skew, and hence make it likely that an oocyst will be produced from the same parents. Large values of alpha tend toward more even densities.

oocyst_distribution
vector specifying the probability distribution of each number of oocysts within the mosquito midgut.

hepatocyte_distribution
vector specifying the probability distribution of the number of infected hepatocytes in a human host. More broadly, this defines the number of independent draws from the oocyst products that make it into the host bloodstream upon a bite from an infectious mosquito.

contig_lengths
vector of lengths (in bp) of each contig.

silent
whether to suppress written messages to the console.